add AssetCard widget to display asset information with profit/loss details and expandable asset list
